requirements.txt
discord.py
requests
tinydb
python-dotenv
aiomysql
mysql